refactor(diff-viewer): 初期 hydrate を分離
